Sofia --- unificación de términos

+++ database/migrations/2020_05_08_091529_create_rooms_table.php
@@ -16,7 +16,7 @@ class CreateRoomsTable extends Migration
         Schema::create('rooms', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('cost')->nullable();
+            $table->float('cost')->nullable();
             $table->string('status')->default();
             $table->timestamps();
         });
